+++ app/Http/Controllers/Tenant/UserController.php
@@ -11,6 +11,7 @@ use App\UserStatus;
 use Illuminate\Validation\Rules;
 use App\Http\Requests\UserStoreRequest;
 use App\Http\Requests\UpdateUserRequest;
+use Illuminate\Support\Facades\Auth;
 
 use Hash;
 
@@ -21,9 +22,12 @@ class UserController extends Controller
      */
     public function index()
     {
+        if (!Auth::user()->hasRole(RolesEnum::SITEMANAGER->value)) {
+            abort(code: 403);
+        }
        // Fetch all tenants from the database
-       $users = User::whereDoesntHave('roles', function ($query) {
-                 $query->where('name', RolesEnum::SITEMANAGER); // Exclude tenant admin role
+       $users = User::whereHas('roles', function ($query) {
+                 $query->where('name', RolesEnum::SITEUSER); // Exclude tenant admin role
              })
              ->orderBy('created_at', 'desc')  // Order by latest
              ->get();      
@@ -35,6 +39,9 @@ class UserController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->hasRole(RolesEnum::SITEMANAGER->value)) {
+            abort(code: 403);
+        }
         return view(view: 'site.user.create');
 
     }
@@ -43,7 +50,10 @@ class UserController extends Controller
      * Store a newly created resource in storage.
      */
     public function store(UserStoreRequest $request)
-    {  
+    { 
+        if (!Auth::user()->hasRole(RolesEnum::SITEMANAGER->value)) {
+            abort(code: 403);
+        } 
         // Create Site Manager User
         $user = User::create([
             'first_name' => $request->first_name,
@@ -53,7 +63,7 @@ class UserController extends Controller
         ]);
         // Assign Role
         $user->assignRole(RolesEnum::SITEUSER);
-        return redirect()->back()->with('success', value: 'User created successfully.');
+        return redirect()->route('users.index')->with('success', value: 'User created successfully.');
     }
 
     /**
@@ -61,6 +71,9 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        if (!Auth::user()->hasRole(RolesEnum::SITEMANAGER->value)) {
+            abort(code: 403);
+        }
        
     }
 
@@ -69,6 +82,9 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        if (!Auth::user()->hasRole(RolesEnum::SITEMANAGER->value)) {
+            abort(code: 403);
+        }
               // Fetch all tenants from the database
         return view('site.user.edit', compact('user'));
     }
@@ -79,6 +95,9 @@ class UserController extends Controller
     public function update(UpdateUserRequest $request, User $user)
     {
 
+        if (!Auth::user()->hasRole(RolesEnum::SITEMANAGER->value)) {
+            abort(code: 403);
+        }
     // Update user details
     $user->first_name = $request->input('first_name');
     $user->last_name = $request->input('last_name');
@@ -100,6 +119,9 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if (!Auth::user()->hasRole(RolesEnum::SITEMANAGER->value)) {
+            abort(code: 403);
+        }
         $user->delete();
 
         // Redirect back with success message
@@ -107,18 +129,21 @@ class UserController extends Controller
     }
     public function toggleStatus(Request $request)
     {
+        if (!Auth::user()->hasRole(RolesEnum::SITEMANAGER->value)) {
+            abort(code: 403);
+        }
         $user = User::findOrFail($request->user_id);
         // Toggle status
-        if ($user->status === UserStatus::ACTIVE->value) {
-            $user->status = UserStatus::DEACTIVE->value;
+        if ($user->status === UserStatus::ACTIVE) {
+            $user->status = UserStatus::DEACTIVE;
         } else {
-            $user->status = UserStatus::ACTIVE->value;
+            $user->status = UserStatus::ACTIVE;
         }
-    
+        // Save the user
+        $user->save();
         return response()->json([
             'success' => true,
-            //'message' => 'Status changed successfully!',
-            'message' => " Status changed successfully!"
+            'message' => 'Status changed successfully!'
         ]);
     }
 }
